Redirect through `rustc`.
